Move IDeploymentStateManager and DeploymentStateSection from Publishing to Pipelines.
Make DeploymentStateSection experimental.
