+++ src/Aspire.Hosting.Azure/Provisioning/Internal/BaseProvisioningContextProvider.cs
@@ -5,7 +5,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Text.RegularExpressions;
-using Aspire.Hosting.Publishing;
+using Aspire.Hosting.Pipelines;
 using Azure;
 using Azure.Core;
 using Azure.ResourceManager.Resources;
