+++ src/Aspire.Hosting/Pipelines/DeploymentStateSection.cs
@@ -1,9 +1,10 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Diagnostics.CodeAnalysis;
 using System.Text.Json.Nodes;
 
-namespace Aspire.Hosting.Publishing;
+namespace Aspire.Hosting.Pipelines;
 
 /// <summary>
 /// Represents a section of deployment state with version tracking for concurrency control.
@@ -14,6 +15,7 @@ namespace Aspire.Hosting.Publishing;
 /// <param name="sectionName">The name of the section.</param>
 /// <param name="data">The JSON data for this section.</param>
 /// <param name="version">The current version of this section.</param>
+[Experimental("ASPIREPIPELINES002", UrlFormat = "https://aka.ms/aspire/diagnostics/{0}")]
 public sealed class DeploymentStateSection(string sectionName, JsonObject? data, long version)
 {
     /// <summary>
