+++ src/Aspire.Hosting/Pipelines/IDeploymentStateManager.cs
@@ -3,7 +3,7 @@
 
 using System.Diagnostics.CodeAnalysis;
 
-namespace Aspire.Hosting.Publishing;
+namespace Aspire.Hosting.Pipelines;
 
 /// <summary>
 /// Provides deployment state management functionality.
