+++ tests/Aspire.Hosting.Tests/Publishing/DeploymentStateManagerTests.cs
@@ -2,17 +2,18 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 #pragma warning disable ASPIREPIPELINES001 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
+#pragma warning disable ASPIREPIPELINES002 // Type is for evaluation purposes only and is subject to change or removal in future updates. Suppress this diagnostic to proceed.
 
 using System.Text.Json.Nodes;
-using Aspire.Hosting.Publishing;
-using Aspire.Hosting.Publishing.Internal;
+using Aspire.Hosting.Pipelines;
+using Aspire.Hosting.Pipelines.Internal;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 
-namespace Aspire.Hosting.Tests.Publishing;
+namespace Aspire.Hosting.Tests.Pipelines;
 
 public class DeploymentStateManagerTests
 {
